added input reader

+++ src/RandomRectangleGenerator.cpp
@@ -6,8 +6,8 @@
 #include "RandomRectangleGenerator.h"
 
 Rectangle RandomRectangleGenerator::generateRandomRect(std::default_random_engine generator) {
-  std::uniform_int_distribution<float> coordinatesDistribution(0,500000);
-  std::uniform_int_distribution<float> heightWidthDistribution(1,100);
+  std::uniform_int_distribution<int> coordinatesDistribution(0,500000);
+  std::uniform_int_distribution<int> heightWidthDistribution(1,100);
   float x1 = coordinatesDistribution(generator);
   float y1 = coordinatesDistribution(generator);
   float height = heightWidthDistribution(generator);
